refact: rename settings historyChart -> chart for consistency.
The historyChart settings contains options related to other charts (flow chart, utxo bubbles), so rename it to indicate its broader scope.
