+++ apps/mobile/locales/en.json
@@ -2476,20 +2476,24 @@
     "features.bip39.description": "Default language used in mnemonic from BIP39 word list",
     "features.bip39.longDescription": "Select the default language for the word list used in mnemonic seed from BIP39 when creating a new account.",
     "features.bip39.title": "Mnemonic Seed Language",
-    "features.charts.historyChart.description": "Control the visible layers and navigation of the transaction history chart",
+    "features.charts.description": "Control the visible layers and navigation of the transaction history, transaction flow and utxo bubble charts",
     "features.charts.historyChart.layers.showAmounts": "Show Amounts",
     "features.charts.historyChart.layers.showFiatAtTxTime": "Show fiat value at the time of the TX",
     "features.charts.historyChart.layers.showFiatOnChart": "Show current fiat value",
     "features.charts.historyChart.layers.showFiatPercentageChange": "Show percentage difference",
     "features.charts.historyChart.layers.showLabels": "Show Labels",
     "features.charts.historyChart.layers.showOutputFields": "Show Output Fields",
-    "features.charts.historyChart.layers.showTransactionFlowChart": "Show Transaction Flow Chart",
     "features.charts.historyChart.layers.showTransactionInfo": "Show Transaction Info",
-    "features.charts.historyChart.layers.showUtxoFlowChart": "Show Utxo Flow Chart",
     "features.charts.historyChart.layers.title": "Layers",
     "features.charts.historyChart.navigation.lockZoomXAxis": "Lock Zoom To X-Axis",
     "features.charts.historyChart.navigation.title": "Navigation",
     "features.charts.historyChart.title": "History Chart",
+    "features.charts.pageTitle": "Chart Settings",
+    "features.charts.title": "Charts",
+    "features.charts.transactionFlowChart.show": "Show Transaction Flow Chart",
+    "features.charts.transactionFlowChart.title": "Transaction Flow Chart",
+    "features.charts.utxoBubbles.show": "Show Utxo Bubbles",
+    "features.charts.utxoBubbles.title": "Utxo Bubbles",
     "features.currencyFormatting.currencyUnit": "Currency unit",
     "features.currencyFormatting.description": "Currency display and formatting",
     "features.currencyFormatting.displayOptions": "Display options",
